+++ tracker/generate/clean_presets.go
@@ -32,9 +32,11 @@ func main() {
 			return nil
 		}
 		tracker.RemoveUnusedUnitParameters(&instr) // remove invalid parameters
-		instr.Name = ""                            // we don't need the names in the preset files as they are derived from the file path
-		instr.NumVoices = 1
-		outData, err := yaml.Marshal(instr)
+		instr2 := sointu.Instrument{               // keep only the relevant fields
+			Comment: instr.Comment,
+			Units:   instr.Units,
+		}
+		outData, err := yaml.Marshal(instr2)
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "could not marshal the preset file %v: %v\n", path, err)
 			return nil